Fix reinstall_2() function to find the latest version of the Paradox Launcher
The reinstall_2() function in UI_logic/MainWindow.py sorts through the
list of paradox launcher versions based on the date they were created
and picks the first element of that sorted list. That means that it
picks the oldest version, not the newest.

This patch reverses the list so it can pick the latest added version.

Also removed duplicated code.